story_creator/requirements.txt
# Django and related packages
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2==2.9.6
Pillow==10.0.0
django-cors-headers==3.14.0
django-jazzmin
dj-database-url
psycopg2